Move Mock class at the send of the test class
